feat(用户标注情况): 用户标注情况的分配审核和查询尚未开发完成
1.查询所有用户功能有错无。
2.分配审核没做完voiceAnnotation更改逻辑
3.查询逻辑尚未完成

+++ ruoyi-admin/src/main/java/com/ruoyi/task/service/impl/TaskServiceImpl.java
@@ -72,6 +72,7 @@ public class TaskServiceImpl implements ITaskService
     @Override
     public List<Task> selectTaskList(Task task)
     {
+
         List<Task> tasks = taskMapper.selectTaskList(task);
         for (Task task1 :tasks){
             if (task1.getAllocationTotal()!=null&&task1.getAllocationTotal()!=0L)
@@ -122,7 +123,7 @@ public class TaskServiceImpl implements ITaskService
                     VoiceAnnotation voiceAnnotation = new VoiceAnnotation();
                     voiceAnnotation.setLabelUser(userId);
                     voiceAnnotation.setUpdateNum(taskAllocationUser.getTaskQuantity());
-                    voiceAnnotation.setTaskId(task_id);
+                    voiceAnnotation.setTaskId((long)task_id);
                     voiceAnnotation.setTaskOwner(user.getUserName());
                     voiceAnnotation.setClazzId(class1Id);
                     voiceAnnotationMapper.updateVoiceAnnotationByClazzId(voiceAnnotation);
